+++ project2-backend/ProfileService/src/main/java/com/revature/repositories/ProfileRepository.java
@@ -12,6 +12,6 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface ProfileRepository extends JpaRepository<Profile, Long> {
-    Profile getProfileByPid(Integer pid);
+    Profile getProfileByPid(Long pid);
     Profile getProfileByUsername(String username);
 }
